+++ src/main/java/org/jsoup/select/Elements.java
@@ -52,16 +52,28 @@ public class Elements extends ArrayList<Element> {
      * @return a deep copy
      */
     @Override
-	public Elements clone() {
+    public Elements clone() {
         Elements clone = new Elements(size());
+        for (Element e : this)
+            clone.add(e.clone());
+        return clone;
+    }
 
-        for(Element e : this)
-    		clone.add(e.clone());
-    	
-    	return clone;
-	}
+    /**
+     Convenience method to get the Elements as a plain ArrayList. This allows modification to the list of elements
+     without modifying the source Document. I.e. whereas calling {@code elements.remove(0)} will remove the element from
+     both the Elements and the DOM, {@code elements.asList().remove(0)} will remove the element from the list only.
+     <p>Each Element is still the same DOM connected Element.</p>
 
-	// attribute methods
+     @return a new ArrayList containing the elements in this list
+     @since 1.19.2
+     @see #Elements(List)
+     */
+    public ArrayList<Element> asList() {
+        return new ArrayList<>(this);
+    }
+
+    // attribute methods
     /**
      Get an attribute value from the first matched element that has the attribute.
      @param attributeKey The attribute key.
@@ -767,7 +779,7 @@ public class Elements extends ArrayList<Element> {
     }
 
     /**
-     Remove the specified Element from this list, and from th DOM
+     Remove the specified Element from this list, and from the DOM.
      * @param o element to be removed from this list, if present
      * @return if this list contained the Element
      * @since 1.17.1
@@ -780,6 +792,26 @@ public class Elements extends ArrayList<Element> {
             remove(index);
             return true;
         }
+    }
+
+    /**
+     * Remove the Element at the specified index in this list, but not from the DOM.
+     * @param index the index of the element to be removed
+     * @return the old element at this index
+     * @since 1.19.2
+     */
+    public Element deselect(int index) {
+        return super.remove(index);
+    }
+
+    /**
+     * Remove the specified Element from this list, but not from the DOM.
+     * @param o element to be removed from this list, if present
+     * @return if this list contained the Element
+     * @since 1.19.2
+     */
+    public boolean deselect(Object o) {
+        return super.remove(o);
     }
 
     /**